requirements.txt
openai
whisper
pydub
python-dotenv
Flask
twilio
gspread[oauth]
requests